Fix oddity in return type of DTMC.getNumTransitions(s) - double not int.
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@8254 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/explicit/DTMC.java
@@ -39,7 +39,7 @@ public interface DTMC extends Model
 	/**
 	 * Get the number of transitions from state s.
 	 */
-	public double getNumTransitions(int s);
+	public int getNumTransitions(int s);
 
 	/**
 	 * Get an iterator over the transitions from state s.
