+++ prism/src/explicit/DTMCUniformisedSimple.java
@@ -198,7 +198,7 @@ public class DTMCUniformisedSimple extends DTMCExplicit
 
 	// Accessors (for DTMC)
 
-	public double getNumTransitions(int s)
+	public int getNumTransitions(int s)
 	{
 		// TODO
 		throw new RuntimeException("Not implemented yet");
